Get a response from Advent of Code

Get the data from the wanted year and day for Advent of Code. Responses will be cached to reduce the number of requests sent to the Advent of Code website.

aoc_get_response(year, day)

Arguments

year

Numeric. Year of the challenge.

day

Numeric. Day of the challenge.

Value

A httr2_response object that will need to be interacted with using httr2's response handling functions.

See also

Examples

if (FALSE) {
resp <- aoc_get_response(2022, 1)
}